requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
google-auth==2.23.2
google-auth-oauthlib==1.1.0
gspread==5.11.3
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==2.0.6
Werkzeug==3.0.0
